+++ app/src/main/java/dllhell/familybudget/presentation/addexpense/AddExpenseView.java
@@ -5,7 +5,7 @@ import com.arellomobile.mvp.MvpView;
 import java.util.Date;
 
 public interface AddExpenseView extends MvpView {
-    void SetSum(long sum);
+    void SetSum(double sum);
 
     void SetDate(Date date);
 }
